backend/requirements.txt
# FastAPI 핵심
fastapi
uvicorn[standard]
sqlalchemy
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
psycopg2-binary
python-dotenv
argon2-cffi
pydantic-settings
email-validator

# 태스크 큐
celery[redis]
redis

# 머신러닝 & 오디오
# 참고: GPU Docker 사용 시 베이스 이미지 의존성을 활용하거나 특정 버전 설치 필요
# PyTorch 표준 설치 (CPU/CUDA 호환 휠)
torch
torchaudio
numpy
metrics
scikit-learn
faiss-cpu
onnx
onnxruntime

# 오디오 처리
librosa
soundfile
mutagen
tqdm
requests
